requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Flask-PyMongo
Flask-SQLAlchemy
Flask-WTF
Werkzeug
WTForms
itsdangerous
Jinja2
MarkupSafe
click
blinker
SQLAlchemy
Mako
greenlet
alembic
pymongo
numpy
pandas
scikit-learn==1.7.1
scipy
joblib
threadpoolctl
python-dateutil
pytz
tzdata
typing_extensions
six
bcrypt
dnspython
idna
colorama
gunicorn
Authlib
email-validator
requests
python-dotenv
# New library for sending emails
Flask-Mail
